requirements.txt
Django>=4.2
django-environ>=0.9.0
django-allauth>=0.55.0
django-ratelimit>=3.0.1
cryptography>=40.0
mysqlclient>=2.1.0
pytest-django>=4.5
pywebpush>=2.0.0
openpyxl>=3.1.2
